requirements.txt
Django>=5.2,<6.0
markdown2>=2.5.0
bleach>=6.2.0
django-taggit>=6.1.0
Pillow>=11.0.0
reportlab>=4.2.0
whitenoise>=6.6.0
gunicorn>=23.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.10
dj-database-url>=2.1.0
